Initial SVN commit of search indexing tool
In current state metadata records with OGC protocol links can be requested and
matching records are harvested as queued_document records. These can then be
traversed by an indexing service that uses each queued_document record to send
a WFS request to a geoserver and index the geometry of the feature type.