#692 #773
XML handling fixes

The XML piped out as a result string was being pretty printed causing
aspects of the portal to break where they weren't trimming whitespace
from element data, have changed the way the XML is generated to prevent
pretty printing.  There was also an error with the adjustment of the
keyword counts where the node was being printed rather than the nodes values.