Fix for WA IVEC and geonetwork changeDate parameter
WA IVEC requires a gml2 response and uses a weird id attribute in the element name
there are a series of changes related to that in this commit.  Also this commit has
a series of changes related to using the geonetwork changeDate to determine if a
metadata record and associated features need to be indexed.